v1.3

+++ models/pg_model.go
@@ -74,7 +74,7 @@ type Pg_ToElement_Mqtt_Obj struct {
 }
 
 type Pg_ToSchedule_Mqtt struct {
-	IDSchedule int `json:"idSchedule"`
-	IDCarta    int `json:"idCarta"`
-	Quantity   int `json:"Quantity"`
+	IDSchedule int64 `json:"idSchedule"`
+	IDCarta    int   `json:"idCarta"`
+	Quantity   int   `json:"Quantity"`
 }
